integration_test/integration_tests/requirements.txt
omotes-sdk-python == 0.0.12
python-dotenv ~= 1.0.0
